src/components/Expense/AddExpenseForm.tsx
import React, { useState } from "react";
import { useContext } from "react";
import { AppContext } from "../../context/AppContext";
import { Expense } from "../../types/types";

const AddExpenseForm = () => {
  // Exercise: Consume the AppContext here
  const {expenses, setExpenses} = useContext(AppContext);

  // Exercise: Create name and cost to state variables
  const[name, setName] = useState("");
  const[cost, setCost] = useState("");

  const onSubmit = (event: React.FormEvent<HTMLFormElement>) => {
    event.preventDefault();

    // Somewhere in your codebase, create a helper function
    const generateUniqueId = () => `${Date.now()}-${Math.floor(Math.random() * 10000)}`;


    // Exercise: Add add new expense to expenses context array
    const newExpense:Expense ={
      // id: crypto.randomUUID(),
      id: generateUniqueId(),
      name, 
      cost: parseFloat(cost),
    };

    setExpenses([...expenses, newExpense]);
    setName("");
    setCost("");

  };

  return (
    <form onSubmit={(event) => onSubmit(event)}>
      <div className="row">
        <div className="col-sm">
          <label htmlFor="name">Name</label>
          <input
            required
            type="text"
            className="form-control"
            id="name"
            value={name}
            onChange={(e) => setName(e.target.value)}
            placeholder="Enter expense name" 
          ></input>
        </div>
        <div className="col-sm">
          <label htmlFor="cost">Cost</label>
          <input
            required
            type="text"
            className="form-control"
            id="cost"
            value={cost}
            onChange={(e) => setCost(e.target.value)}
            placeholder="Enter amount" 
          ></input>
        </div>
        <div className="col-sm">
          <button type="submit" className="btn btn-primary mt-3">
            Save
          </button>
        </div>
      </div>
    </form>
  );
};

export default AddExpenseForm;
